+++ front-end/src/pages/CardDetails/CardDetails.js
@@ -1,9 +1,10 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom'
-import * as backEndApi from '../api/BackEndApi';
-import Card from '../components/Card';
-import Footer from '../components/Footer';
-import Header from '../components/Header';
+import * as backEndApi from '../../api/BackEndApi';
+import Card from '../../components/Card/Card';
+import Footer from '../../components/Footer/Footer';
+import Header from '../../components/Header/Header';
+import styles from './styles.module.css';
 
 const CardDetails = () => {
   const [card, setCard] = useState({});
@@ -20,11 +21,11 @@ const CardDetails = () => {
   if (!card) return 'Card not found';
 
   return (
-    <>
+    <section className={styles["page-container"]}>
       <Header />
       <Card content={card} />
       <Footer />
-    </>
+    </section>
   );
 };